requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
pillow
numpy
pydantic
